refactor: move create questionnaire functionality out
refs: BLAIS5-3218
